Remove casting away of volatile and explicit call to seL4_Yield.
This commit removes some bad style in this example. In particular, calling a
seL4 system call directly from a CAmkES component is not intended. Additionally
this example cast away the volatility of the dataport pointer. From what I can
see, the only reason the compiler did not optimise away the repeated read of
this pointer was because the loop contained a system call with a contained
unrestricted memory clobber.

Related to JIRA CAMKES-436